Option to generate last events overall results as index.html
Option "useResultsForHome" can now be used in initialstate.js and if set to true it will generate last events overall driver results as index.html instead of home template or Standings.

Quickfix which should be re-implemented in a better and more logical way

+++ src/output/html.js
@@ -190,7 +190,7 @@ const writeStandingsHTML = (division, type, links) => {
     `./${outputPath}/website/${division.divisionName}-${type}-standings.html`,
     out
   );
-  if (!fs.existsSync(`./${outputPath}/website/index.html`)) {
+  if (!fs.existsSync(`./${outputPath}/website/index.html`) && leagueRef.league.useStandingsForHome) {
     fs.writeFileSync(`./${outputPath}/website/index.html`, out);
   }
 };
@@ -376,6 +376,9 @@ const writeDriverResultsHTML = (event, division, links, eventIndex) => {
     `./${outputPath}/website/${division.divisionName}-${eventIndex}-driver-results.html`,
     out
   );
+  if (division.divisionName == "overall" && leagueRef.league.useResultsForHome) {
+  fs.writeFileSync(`./${outputPath}/website/index.html`, out);
+  }
 };
 
 const writeHTMLOutputForDivision = (division, links) => {
